Moved blender files around, changes to make sure the legacy IO worked
